microservices/ai_layer/requirements.txt
flask
python-dotenv
langgraph
langchain
langchain-openai
langchain-community
langchain-neo4j